+++ pkg/rid/server/isa_handler.go
@@ -75,6 +75,7 @@ func (s *Server) CreateIdentificationServiceArea(
 		ID:    id,
 		URL:   params.GetFlightsUrl(),
 		Owner: owner,
+		Writer: s.Locality,
 	}
 
 	if err := isa.SetExtents(params.Extents); err != nil {
